Refacto: Moved the datamodel of javaumltranslator into a separate project
