+++ datamodel/package.json
@@ -5,11 +5,12 @@
     "main": "dist/index.js",
     "type": "module",
     "scripts": {
+        "build": "tsc",
         "lint:check": "eslint .",
         "lint:write": "eslint . --fix",
         "format:write": "prettier --write .",
         "format:check": "prettier --check .",
-        "test": "npm test"
+        "test": "vitest run --coverage --config vitest.config.ts"
     },
     "repository": {
         "type": "git",
@@ -26,9 +27,12 @@
     },
     "homepage": "https://github.com/ThalesGroup/lumber#readme",
     "devDependencies": {
-        "@types/node": "^25.5.0",
+        "@types/node": "^25.5.2",
+        "@vitest/coverage-v8": "^4.1.2",
+        "esbuild": "^0.28.0",
         "eslint": "^10.1.0",
         "prettier": "^3.8.1",
-        "typescript": "^6.0.2"
+        "typescript": "^6.0.2",
+        "vitest": "^4.1.2"
     }
 }
