+++ datamodel/src/datamodel/Class.ts
