+++ datamodel/src/datamodel/Enum.ts
