+++ datamodel/src/datamodel/EnumProperty.ts
