+++ datamodel/src/datamodel/index.ts
@@ -1,7 +1,7 @@
+export * from '../generator/DataModelTranslator';
 export * from './Association';
 export * from './Attribute';
 export * from './Class';
-export * from './DataModelTranslator';
 export * from './Enum';
 export * from './EnumProperty';
 export * from './Interface';
