+++ datamodel/vitest.config.ts
@@ -1,0 +1,12 @@
+import { defineConfig } from 'vitest/config';
+
+export default defineConfig({
+    test: {
+        globals: true,
+        environment: 'node',
+        coverage: {
+            reporter: ['text', 'json', 'html']
+        },
+        include: ['src/**/*.test.ts']
+    }
+});
